Doctor images

+++ api/src/controllers/usersControllers.js
@@ -1,44 +1,25 @@
 const { db } = require('../firebase');
 
 //  --- Sign up ---
-const signUpUser = async ({ uid }) => {
+const signUpUser = async ({ email, uid }) => {
     try {
-        const newUser = await db.collection('users').doc(uid).add({
-            name: '',
-            lastName: '',
-            userId: '',
-            date: '',
-            photo: {}
-        });
+      const userRef = db.collection('users').doc(uid);
 
-        return newUser;
+      await userRef.set({
+        email,
+        name: '',
+        id: '',
+        photo: {},
+        dates: []
+      });
+
+      return userRef.id; // El ID del documento es igual al UID del usuario
     } catch (error) {
-        throw new Error(error)
+      throw new Error(error);
     }
-};
+  };
 
-// --- Login ---
-const logInUser = async (email, password) => {
-    try {
-        const userData = await db.collection('users').where('email', '==', email).get();
-        const user = [];
-        userData.forEach((us) => {
-            user.push({
-                id: us.id,
-                ...us.data()
-            })
-        })
-        if (user.length < 1) throw new Error('Mail not registered');
-        if (user[0].password !== password) throw new Error('Unvalid mail or password');
-        else return true;
-    } catch (error) {
-        console.log(error);
-        throw new Error(error)
-    }
-};
-
-
-//? --- Update user ---
+// --- Update user ---
 
 const createUser = async ({ name, email, password, personalId, location, enable, photo }) => {
     try {
@@ -70,15 +51,11 @@ const bringUserById = async (id) => {
 
     try {
         const userData = await db.collection('users').doc(id).get();
-       
-        if (userData.empty) {
-            throw new Error(`No user matched with UID: ${id}`);
-        }
         const user = {
             id: userData.id,
             ...userData.data()
         };
-
+        if(!user.email) throw new Error(`No user matched with UID: ${id}`);
         return user;
 
     } catch (error) {
@@ -87,10 +64,17 @@ const bringUserById = async (id) => {
 };
 
 // --- Delete an user from data base ---
+
 const deleteUser = async (id) => {
     try {
-        const deletedUser = await db.collection('users').doc(id).delete();
-        return deletedUser;
+        const userRef = await db.collection('users').doc(id).get();
+        const user = {
+            id: userRef.id,
+            ...userRef.data()
+        }
+        if(!user.email) throw new Error(`No user matched with UID: ${id}`)
+        await db.collection('users').doc(id).delete()
+        return user
         // Falta tirar error al no encontrar usuario
     } catch (error) {
         console.log(error);
@@ -98,12 +82,47 @@ const deleteUser = async (id) => {
     }
 }
 
+// --- Enable an user ---
+
+const enableUser = async (id) => {
+    try {
+        const enabledUser = await db.collection('users').doc(id).get();
+        const user = {
+            id: enabledUser.id,
+            ...enabledUser.data()
+        }
+        if(!user.email) throw new Error(`user with id ${id} not found`);
+        if(user.enable) throw new Error(`user with ID ${id} already enabled`);
+
+        await db.collection('users').doc(id).update({
+            enable: true
+        });
+        user.enable = true;
+        return user;
+
+    } catch (error) {
+        throw new Error(error)
+    }
+};
+
 // --- Disable an user from data base --- 
+
 const disableUser = async (id) => {
     try {
-        return await db.collection('users').doc(id).update({
+        const disabledUser = await db.collection('users').doc(id).get();
+        const user = {
+            id: disabledUser.id,
+            ...disabledUser.data()
+        }
+        if(!user.email) throw new Error(`user with ID ${id} not found`);
+        if(!user.enable) throw new Error(`user with ID ${id} already disabled`);
+
+        await db.collection('users').doc(id).update({
             enable: false
         })
+        user.enable = false;
+        return user;
+
     } catch (error) {
         throw new Error(error)
     }
@@ -126,15 +145,4 @@ const updateUser = async ({ name, lastName, photo, userId, uid, date }) => {
     }
 };
 
-// // --- Forgot Password ---
-// const newPassword = async (email) => {
-//     try {
-//         return await db.collection('users')
-//     } catch (error) {
-
-//     }
-// }
-
-
-
-module.exports = { createUser, bringUserById, deleteUser, disableUser, signUpUser, logInUser, updateUser }
+module.exports = { createUser, bringUserById, deleteUser, disableUser, signUpUser, updateUser, enableUser }
